Create the database once per class not once per test
Each test consisted of 30 seconds of database creation and 1
second of testing. Creating and dropping the database each
test meant that the entire test module was taking 150 seconds
(on my laptop). The tests do not modify the database so it is
safe to create it once and drop it once. With this change the
entire test module runs in 32 seconds.