Add generation of column lists for datalinker
To support datalinker endpoints that return selected lists of
columns from tables, we want to communicate to datalinker the
list of columns with specific Felis tags in various tables.

Add a Python script that generates a YAML file of columns tagged
with tap:principal, and create a zip file as a release artifact
that can be downloaded by datalinker when it starts.

Add a placeholder columns-minimal.yaml file that will be maintained
by hand for now.