Add new obs_title and lsst_ccdvisitid columns
obs_title is a standard-but-optional column and is planned for use
in labeling images in the Portal display.

lsst_ccdvisitid is a combination of lsst_visit and lsst_detector meant
for use in accelerating JOINs with tables with a "ccdvisitid" column.

Also: switch order of lsst_filter and lsst_visit to a more logical form,
putting filter between the tract-patch columns and the visit-detector
columns, since it applies to both systems.