Fix issues with single quote usage in DP0.1 schema
The string "Source''s" is changed to remove improper usage of two
single quotes. The description fields containing this string are now
enclosed in double quotes so that the single quotes do not need to be
escaped. The string "''regauassianization" is changed to be enclosed
in double quotes. The RSP was checked to confirm that unescaped double
quotes are displayed correctly for other fields where they are
currently used.