Remove length fields on timestamp columns
The length field is disallowed on timestamp columns, since this
datatype is not considered a sized type by Felis where it would be used
in SQL generation. These values are really precision rather than
length, which is not currently supported in the basic type system.
The overrides for MySQL were left in place so that precision could be
specified if necessary.