Create service descriptor for hoverdrive endpoints
Hoverdrive provides two endpoints for getting documentation links about
tables and columns respectively. These endpoints are described in
https://sqr-086.lsst.io.

Both endpoints provide an optional mode triggered by a ?redirect=true
query parameter where the client is redirected to the most-relevant
documentation URL. In this case, only a single column or table can be
specified by the ?table and ?column parameters. I'm not sure how to
encode this logic in the service descriptor.