DM-51604: Modify foreign-key descriptions to clarify 1:N relationships
Note that this "reverses the sense" of all the affected messages.
This is definitely the phrasing that I want, because TAP_SCHEMA
clients basically treat these foreign-key links as bidirectional,
and the 1:N nature is important to convey to the user.

I realize there is some awkwardness about this from a formal RDBMS
modeling perspective, but this is the right UX for science users
given the existing tool environment.